Refactor folder layout; Set path aliases; Remove unused imports; Bump version
