+++ src/commands/_CommandList.ts
@@ -1,19 +1,19 @@
 // Keeps track of all commands as an array
 
-import { CommandInterface } from "../interfaces/Command";
+import { CommandInterface } from "interfaces/Command";
 import { animAvatarTest } from "./animAvatarTest";
-import { anime } from "./anime";
-import { avatar } from "./avatar";
-import { caption } from "./caption";
-import { confess } from "./confession";
-import { guildSettings } from "./guildSettings";
-import { movie } from "./movie";
-import { petPet } from "./petpet";
-import { poke } from "./poke";
-import { ship } from "./ship";
-import { starboard } from "./starboard";
-import { stats } from "./stats";
-import { wordleStats } from "./wordleStats";
+import { anime } from "./slashCommands/anime";
+import { avatar } from "./slashCommands/avatar";
+import { caption } from "./contextCommands/caption";
+import { confess } from "./slashCommands/confession";
+import { guildSettings } from "./slashCommands/guildSettings";
+import { movie } from "./slashCommands/movie";
+import { petPet } from "./slashCommands/petpet";
+import { poke } from "./slashCommands/poke";
+import { ship } from "./slashCommands/ship";
+import { starboard } from "./slashCommands/starboard";
+import { stats } from "./slashCommands/stats";
+import { wordleStats } from "./slashCommands/wordleStats";
 
 // List of all commands registered to the bot
 // These commands are registered to Discord when the bot starts
