+++ src/commands/slashCommands/starboard.ts
@@ -1,12 +1,12 @@
-import { Channel, CommandInteraction, Embed, Message, PermissionsBitField, SlashCommandBuilder } from "discord.js";
-import { CommandInterface } from "../interfaces/Command";
-import { getGuildDataByGuildID, removeStoredStarboardPost, setStarboardChannel as setChannel, setStarboardEmojis as setEmojis, setStarboardThreshold as setThreshold, update } from "../database/guildData";
-import { hasPermissions } from "../utils/userUtils";
+import { Channel, CommandInteraction, Message, PermissionsBitField, SlashCommandBuilder } from "discord.js";
+import { CommandInterface } from "interfaces/Command";
+import { getGuildDataByGuildID, removeStoredStarboardPost, setStarboardChannel as setChannel, setStarboardEmojis as setEmojis, setStarboardThreshold as setThreshold, update } from "database/guildData";
+import { hasPermissions } from "utils/userUtils";
 import { EmbedBuilder } from "@discordjs/builders";
-import { BOT } from "../index";
-import { truncateString } from "../utils/utils";
-import { StarboardLeaderboard, StarboardPost } from "../database/models/guildModel";
-import { commandNotImplemented } from "../utils/commandUtils";
+import { BOT } from "index";
+import { truncateString } from "utils/utils";
+import { StarboardLeaderboard, StarboardPost } from "database/models/guildModel";
+import { commandNotImplemented } from "utils/commandUtils";
 
 const setStarboardChannel = async (interaction: any): Promise<string> => {
     // Check if user has permissions
