+++ src/config/config.json
@@ -1,6 +1,6 @@
 {
-    "version": "1.3.3",
-    "statuses": [ "Back and better than ever!", "Now supports Tradle!", "your mom", "Check out my confession command!", "I do starboard too 👀", "/ship", ":)" ],
+    "version": "1.4.0",
+    "statuses": [ "Back and better than ever!", ":)" ],
     "starboardConfig": {
         "fetchLimit": 100,
         "defaultEmoji": "⭐",
