+++ src/database/connectDatabase.ts
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import userModel from "./models/userModel";
+import userModel from "@models/userModel";
 
 
 // Makes mongoose connection to MongoDB Atlas database
