+++ src/database/wordleData.ts
@@ -1,6 +1,4 @@
-
-import { Aggregate } from "mongoose";
-import wordleModel, { WordleDataInterface, createNewWordleData } from "./models/wordleModel";
+import wordleModel, { WordleDataInterface, createNewWordleData } from "@models/wordleModel";
